% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigDataFunctions.R
\name{bd_dim}
\alias{bd_dim}
\title{Get Dimensions of file
Description: This function returns the number of rows and columns in a data set without actually loading the file. This helps the user decide if all or part of a file should be loaded.}
\usage{
bd_dim(x)
}
\description{
Get Dimensions of file
Description: This function returns the number of rows and columns in a data set without actually loading the file. This helps the user decide if all or part of a file should be loaded.
}
