% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigDataFunctions.R
\name{bd_minMaxMean}
\alias{bd_minMaxMean}
\title{Get the min max mean standard dev normality and quartiles of each column}
\usage{
bd_minMaxMean(x, sep, colnms, rownms)
}
\description{
Get the min max mean standard dev normality and quartiles of each column
}
