% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigDataFunctions.R
\name{bd_orderColNames}
\alias{bd_orderColNames}
\title{Order by colnames
Description: Order maxtrix x by matrix y's columns}
\usage{
bd_orderColNames(x, y)
}
\description{
Order by colnames
Description: Order maxtrix x by matrix y's columns
}
