% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigDataFunctions.R
\name{bd_orderRowNames}
\alias{bd_orderRowNames}
\title{Order by Rownames
Description: Order matrix rownames by matrix y's rowname order}
\usage{
bd_orderRowNames(x, y)
}
\description{
Order by Rownames
Description: Order matrix rownames by matrix y's rowname order
}
