% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigDataFunctions.R
\name{bd_spacewrite}
\alias{bd_spacewrite}
\title{Simplified write file function
Description: This function will write a table of data to file using space delimitation without quotes}
\usage{
bd_spacewrite(x, name, rownms, colnms)
}
\description{
Simplified write file function
Description: This function will write a table of data to file using space delimitation without quotes
}
