\name{bigdata-package}
\alias{bigdata-package}
\docType{package}
\title{
Big Data Analytics
}
\description{
a collection of scalable methods for large-scale data analysis.
}
\details{
\tabular{ll}{
Package: \tab bigdata\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-04-06\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Han Liu, Tuo Zhao
Maintainers: Han Liu <hanliu@cs.jhu.edu>
}
\references{
1. Han Liu, Kathryn Roeder and Larry Wasserman. Stability Approach to Regularization Selection (StARS) for High Dimensional Graphical Models. \emph{Advances in Neural Information Processing Systems}(NIPS), 2010.\cr
}
\seealso{
\code{\link{lasso.stars}}
}