% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigData.R
\name{bigData}
\alias{bigData}
\title{bigData Package}
\description{
Contains functions to pre-process large data files based on conditions
Also merging data sets before, after, or during data loading into memory.
Lastly, it has some nice parsing functions once dat is loaded.
}
\author{
Jared Streich \email{streich.jared@gmail.com}
}
