\name{plot.stars}
\alias{plot.stars}
\title{
Plot function for S3 class "stars"
}
\description{
Visualize the solution path and plot the optimal solution by model selection
}
\usage{
\method{plot}{stars}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"stars"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Han Liu and Tuo Zhao\cr
Maintainers: Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{lasso.stars}}
}