\name{print.stars}
\alias{print.stars}
\title{
Print function for S3 class "stars"
}
\description{
Print the information about the solution path length and the degree of freedom's along the solution path.
}
\usage{
\method{print}{stars}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"stars"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Han Liu and Tuo Zhao\cr
Maintainers: Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{lasso.stars}}
}