% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigDataFunctions.R
\name{bd_2MatrixColMatch}
\alias{bd_2MatrixColMatch}
\title{get x from y based on two specific columns}
\usage{
bd_2MatrixColMatch(x, xc, y, yc)
}
\description{
get x from y based on two specific columns
}
