% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigDataFunctions.R
\name{bd_nameMatch}
\alias{bd_nameMatch}
\title{Intersection Name Match- Name-based Cropping
Description: Crop matrix x by matching rownames in matrix y}
\usage{
bd_nameMatch(x, y)
}
\description{
Intersection Name Match- Name-based Cropping
Description: Crop matrix x by matching rownames in matrix y
}
