% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigDataFunctions.R
\name{bd_thresh}
\alias{bd_thresh}
\title{Read in values smaller, smaller-equal, exactly, greater-equal, greater than in specific column}
\usage{
bd_thresh(x, column, colnms, rownms, threshold, sep, method)
}
\description{
Read in values smaller, smaller-equal, exactly, greater-equal, greater than in specific column
}
